nvidia-stock-predictor/requirements.txt
yfinance
newsapi-python
vaderSentiment
pandas
numpy
matplotlib
seaborn
scikit-learn
statsmodels
pmdarima
tensorflow
keras
transformers
sentencepiece

